pi/requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
PyYAML==6.0.1
tensorflow==2.13.0
pyserial==3.5